formating numbers
git-svn-id: http://emr.moh.gov.rw/svn/modules/moh-billing/branches/1.6.x@3193 6101c9c6-cf70-4d46-9f71-e9b75042c301

+++ src/org/openmrs/module/mohbilling/businesslogic/FileExporter.java
@@ -772,7 +772,7 @@ public class FileExporter {
 			PdfPTable table = new PdfPTable(1);
 			 for (PatientServiceBill psb : pb.getBillItems()) {
 				 number++;
-				 Double serviceCost = psb.getUnitPrice().doubleValue()*psb.getQuantity();
+				 Double serviceCost = psb.getUnitPrice().doubleValue()*psb.getQuantity().doubleValue();
 			     c = new PdfPCell(font.process(number+")"+psb.getService().getFacilityServicePrice().getName()+" "+ReportsUtil.roundTwoDecimals(psb.getUnitPrice().doubleValue()*getPatientRate(pb)/100)+" x "+psb.getQuantity()+" = "+serviceCost*getPatientRate(pb)/100+"\n"));
 					
 			}
